[FAT-153] Remove TestRail integration
