+++ acquisitions/src/test/java/org/folio/EbsconetApiTest.java
@@ -1,24 +1,20 @@
 package org.folio;
 
-import org.folio.testrail.AbstractTestRailIntegrationTest;
-import org.folio.testrail.config.TestModuleConfiguration;
-import org.folio.testrail.services.TestRailIntegrationService;
+import org.folio.test.TestBase;
+import org.folio.test.config.TestModuleConfiguration;
+import org.folio.test.services.TestIntegrationService;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 
-public class EbsconetApiTest extends AbstractTestRailIntegrationTest {
+public class EbsconetApiTest extends TestBase {
 
   // default module settings
   private static final String TEST_BASE_PATH = "classpath:domain/mod-ebsconet/features/";
-  private static final String TEST_SUITE_NAME = "mod-ebsconet";
-  // TEST_SUITE_ID and TEST_SECTION_ID are obtained from TestRail
-  private static final long TEST_SUITE_ID = 742L;
-  private static final long TEST_SECTION_ID = 11516L;
 
   public EbsconetApiTest() {
-    super(new TestRailIntegrationService(
-      new TestModuleConfiguration(TEST_BASE_PATH, TEST_SUITE_NAME, TEST_SUITE_ID, TEST_SECTION_ID)));
+    super(new TestIntegrationService(
+      new TestModuleConfiguration(TEST_BASE_PATH)));
   }
 
   @Test
