+++ acquisitions/src/test/java/org/folio/FinanceApiTest.java
@@ -1,24 +1,20 @@
 package org.folio;
 
-import org.folio.testrail.AbstractTestRailIntegrationTest;
-import org.folio.testrail.config.TestModuleConfiguration;
-import org.folio.testrail.services.TestRailIntegrationService;
+import org.folio.test.TestBase;
+import org.folio.test.config.TestModuleConfiguration;
+import org.folio.test.services.TestIntegrationService;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 
-public class FinanceApiTest extends AbstractTestRailIntegrationTest {
+public class FinanceApiTest extends TestBase {
 
   // default module settings
   private static final String TEST_BASE_PATH = "classpath:domain/mod-finance/features/";
-  private static final String TEST_SUITE_NAME = "mod-finance";
-  private static final long TEST_SECTION_ID = 3347L;
-  // TODO: make TEST_SUITE_ID different for each module
-  private static final long TEST_SUITE_ID = 161L;
 
   public FinanceApiTest() {
-    super(new TestRailIntegrationService(
-        new TestModuleConfiguration(TEST_BASE_PATH, TEST_SUITE_NAME, TEST_SUITE_ID, TEST_SECTION_ID)));
+    super(new TestIntegrationService(
+        new TestModuleConfiguration(TEST_BASE_PATH)));
   }
 
   @Test
