+++ acquisitions/src/test/java/org/folio/OrdersApiTest.java
@@ -1,25 +1,21 @@
 package org.folio;
 
-import org.folio.testrail.AbstractTestRailIntegrationTest;
-import org.folio.testrail.config.TestModuleConfiguration;
-import org.folio.testrail.services.TestRailIntegrationService;
+import org.folio.test.TestBase;
+import org.folio.test.config.TestModuleConfiguration;
+import org.folio.test.services.TestIntegrationService;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
 
-public class OrdersApiTest extends AbstractTestRailIntegrationTest {
+public class OrdersApiTest extends TestBase {
 
   // default module settings
   private static final String TEST_BASE_PATH = "classpath:domain/mod-orders/features/";
-  private static final String TEST_SUITE_NAME = "mod-orders";
-  private static final long TEST_SECTION_ID = 3337L;
-  // TODO: make TEST_SUITE_ID different for each module
-  private static final long TEST_SUITE_ID = 159L;
 
   public OrdersApiTest() {
-    super(new TestRailIntegrationService(
-        new TestModuleConfiguration(TEST_BASE_PATH, TEST_SUITE_NAME, TEST_SUITE_ID, TEST_SECTION_ID)));
+    super(new TestIntegrationService(
+        new TestModuleConfiguration(TEST_BASE_PATH)));
   }
   @Test
   void deleteOpenedOrderAndOrderLines() {
