+++ mod-data-export/src/test/java/org/folio/ModDataExportApiTest.java
@@ -1,8 +1,8 @@
 package org.folio;
 
-import org.folio.testrail.AbstractTestRailIntegrationTest;
-import org.folio.testrail.config.TestModuleConfiguration;
-import org.folio.testrail.services.TestRailIntegrationService;
+import org.folio.test.TestBase;
+import org.folio.test.config.TestModuleConfiguration;
+import org.folio.test.services.TestIntegrationService;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.MethodOrderer;
@@ -11,16 +11,13 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.TestMethodOrder;
 
 @TestMethodOrder(MethodOrderer.OrderAnnotation.class)
-class ModDataExportApiTest extends AbstractTestRailIntegrationTest {
+class ModDataExportApiTest extends TestBase {
 
-    private static final long TEST_SUITE_ID = 114L;
-    private static final long TEST_SECTION_ID = 4517L;
-    private static final String TEST_SUITE_NAME = "mod-data-export";
     private static final String TEST_BASE_PATH = "classpath:domain/dataexport/features/";
 
     public ModDataExportApiTest() {
-        super(new TestRailIntegrationService(
-                new TestModuleConfiguration(TEST_BASE_PATH, TEST_SUITE_NAME, TEST_SUITE_ID, TEST_SECTION_ID)));
+        super(new TestIntegrationService(
+                new TestModuleConfiguration(TEST_BASE_PATH)));
     }
 
     @Test
