+++ quick-marc/src/test/java/org/folio/QuickMarcApiTest.java
@@ -1,23 +1,20 @@
 package org.folio;
 
-import org.folio.testrail.AbstractTestRailIntegrationTest;
-import org.folio.testrail.config.TestModuleConfiguration;
-import org.folio.testrail.services.TestRailIntegrationService;
+import org.folio.test.TestBase;
+import org.folio.test.config.TestModuleConfiguration;
+import org.folio.test.services.TestIntegrationService;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 
-public class QuickMarcApiTest extends AbstractTestRailIntegrationTest {
+public class QuickMarcApiTest extends TestBase {
 
     private static final String TEST_BASE_PATH = "classpath:domain/mod-quick-marc/features/";
-    private static final String TEST_SUITE_NAME = "mod-quick-marc";
-    private static final long TEST_SECTION_ID = 1326L;
-    private static final long TEST_SUITE_ID = 48L;
 
     public QuickMarcApiTest() {
-        super(new TestRailIntegrationService(
-            new TestModuleConfiguration(TEST_BASE_PATH, TEST_SUITE_NAME, TEST_SUITE_ID, TEST_SECTION_ID)));
+        super(new TestIntegrationService(
+            new TestModuleConfiguration(TEST_BASE_PATH)));
     }
 
     @Test
