+++ testrail-integration/src/main/java/org/folio/test/config/TestModuleConfiguration.java
@@ -1,0 +1,15 @@
+package org.folio.test.config;
+
+public class TestModuleConfiguration {
+
+  private final String basePath;
+
+  public TestModuleConfiguration(String basePath) {
+    this.basePath = basePath;
+  }
+
+  public String getBasePath() {
+    return basePath;
+  }
+
+}
