--- testrail-integration/src/main/java/org/folio/testrail/api/TestRailException.java
@@ -1,10 +1,0 @@
-package org.folio.testrail.api;
-
-public class TestRailException extends Exception {
-
-  private static final String TESTRAIL_ERROR_MESSAGE = "TestRail API return HTTP ";
-
-  public TestRailException(String message) {
-    super(TESTRAIL_ERROR_MESSAGE + message);
-  }
-}
