--- testrail-integration/src/main/java/org/folio/testrail/models/TestRailStatus.java
@@ -1,18 +1,0 @@
-package org.folio.testrail.models;
-
-public enum TestRailStatus {
-  PASSED(1),
-  BLOCKED(2),
-  UNTESTED(3),
-  RETEST(4), FAILED(5);
-
-  private int statusId;
-
-  TestRailStatus(int i) {
-    this.statusId = i;
-  }
-
-  public int getStatusId() {
-    return statusId;
-  }
-}
